Refactor for testability
This is a pretty big patch that almost completely rewrites the app.
The main purpose of the refactor is to make the core functionality
testable, which can now be done by executing `test.c`.

As part of this change the core input sending logic has been rebuilt
to now send keys that have both virtual key codes and scan codes.

In general it seems that virtual key codes work better in most cases;
hopefully the scan code being embedded in the send input signal will
also make the inputs work in games that otherwise ignore virtual key
codes.