Make console toggle compatible with win11 terminal
Windows 11 uses a new Terminal app as the default console. This console
does not support the usual show/hide tricks that worked with the ol
command prompt. As a result w11 users report that the console window
sticks around after launching dual-key-remap.

A simple fix for now is to compile to GUI app and manually create
or destroy a console window as needed.

Ideally we don't create a console window unless we need it. Right now
we create a console no matter what and hide if it's NOT needed.